+++ tests/analyse/test_modelled_obs.py
@@ -307,3 +307,15 @@ def test_modelled_obs_co2_consistency(model_scenario_co2_dummy, model_scenario_p
     combined_paris = model_scenario_paris_co2_dummy.footprints_data_merge()
 
     xr.testing.assert_equal(combined.mf_mod_high_res, combined_paris.mf_mod_high_res)
+
+
+# TODO: this test could go elsewhere?
+def test_model_modelled_obs_co2_multisector(model_scenario_co2_dummy, flux_co2_dummy):
+    """Test footprints_data_merge with multisector return options"""
+    model_scenario_co2_dummy.add_flux(species="co2", flux={"TESTSOURCE2": flux_co2_dummy})
+    combined_dataset = model_scenario_co2_dummy.footprints_data_merge(calc_fp_x_flux=True, split_by_sectors=True)
+    print(combined_dataset)
+    print(combined_dataset.data_vars)
+    assert all(dv in combined_dataset for dv in ["mf_mod_high_res", "fp_x_flux", "mf_mod_high_res_sectoral", "fp_x_flux_sectoral"])
+
+    assert all(combined_dataset.source.values == ["TESTSOURCE", "TESTSOURCE2"])